% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadRouteR-package.R
\docType{package}
\name{roadRouteR}
\alias{roadRouteR}
\alias{roadRouteR-package}
\title{roadRouteR: Route Calculation and Mapping with OSRM + OpenStreetMap}
\description{
A lightweight wrapper for calculating and visualizing realistic road routes using OSRM and OpenStreetMap.
}
\section{Functions}{

- [getRoute()]: Calculate and optionally plot a road route between two places.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ridout-m/roadRouteR}
  \item Report bugs at \url{https://github.com/ridout-m/roadRouteR/issues}
}

}
\keyword{package}
