% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadRouteR-package.R
\docType{package}
\name{roadRouteR}
\alias{roadRouteR}
\alias{roadRouteR-package}
\title{roadRouteR: Route Calculation and Mapping}
\description{
A lightweight wrapper for calculating and visualizing realistic road routes using OSRM and OpenStreetMap.
}
\section{Functions}{

\itemize{
 \item \code{getRoute()}:  Calculate and optionally plot a road route between two places
}
}

\section{Dependencies}{

\itemize{
  \item \code{osrm}:  Communicates with the OSRM backend
  \item \code{tidygeocoder}:  Converts place names/postcodes to coordinates using OSM
  \item \code{sf}:  Handles spatial data and geometry
  \item \code{leaflet}:  Plots interactive web maps
  \item \code{geosphere}:  Detects route jumps or broken segments
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ridout-m/roadRouteR}
  \item Report bugs at \url{https://github.com/ridout-m/roadRouteR/issues}
}

}
\author{
Mackenzie Ridout \url{https://github.com/ridout-m}
}
\keyword{package}
